feat(ruby): port scanner to C API and pin to tree-sitter 0.26.3
Replace the vendored C++ scanner.cc with upstream's C scanner plus the
Semgrep ellipsis extension, and move ruby onto the 0.26.3 language lists.
