+++ lang/language-variants-0.22.6
@@ -26,7 +26,6 @@ ocaml
 promql
 proto
 ql
-ruby
 rust
 scala
 sqlite
