+++ lang/language-variants-0.26.3
@@ -5,3 +5,4 @@ php
 php-only
 powershell
 python
+ruby
