+++ lang/languages-0.22.6
@@ -28,7 +28,6 @@ promql
 proto
 ql
 requirements
-ruby
 rust
 scala
 sqlite
