+++ lang/languages-0.26.3
@@ -4,3 +4,4 @@ gosu
 php
 powershell
 python
+ruby
